requirements.txt
amqp==2.6.0
asn1crypto==0.24.0
billiard==3.6.3.0
celery==4.4.3
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
fail2ban==0.10.2
Flask==1.1.2
Flask-Cors==3.0.8
httplib2==0.9.2
idna==2.6
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.9
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
macaroonbakery==1.1.3
MarkupSafe==1.1.1
netifaces==0.10.4
numpy==1.18.4
oauth==1.0.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.13.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
usb-creator==0.3.3
vine==1.3.0
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==1.0.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
